+++ src/Graph.java
@@ -3,28 +3,28 @@ import java.util.List;
 import java.util.Map;
 
 public class Graph {
-	private BigInteger n; //Nombre de Noeud
-	private BigInteger nbEdges; //Nombre d'Edge
+	private int n; //Nombre de Noeud
+	private int nbEdges; //Nombre d'Edge
 	private List<Edge> edgeList; // List des edges
-	private Map<BigInteger, Integer> degree; // List des edges
+	private Map<Integer, Integer> degree; // List des edges
 	
-	public Graph(BigInteger n, BigInteger nbEdges, List<Edge> edgeList, Map<BigInteger, Integer> degree) {
+	public Graph(int n, int nbEdges, List<Edge> edgeList, Map<Integer, Integer> degree) {
 		this.n = n;
 		this.nbEdges = nbEdges;
 		this.edgeList = edgeList;
 		this.degree = degree;
 	}
 	
-	public BigInteger getN() {
+	public int getN() {
 		return n;
 	}
-	public void setN(BigInteger n) {
+	public void setN(int n) {
 		this.n = n;
 	}
-	public BigInteger getNbEdges() {
+	public int getNbEdges() {
 		return nbEdges;
 	}
-	public void setNbEdges(BigInteger nbEdges) {
+	public void setNbEdges(int nbEdges) {
 		this.nbEdges = nbEdges;
 	}
 	public List<Edge> getEdgeList() {
@@ -33,10 +33,10 @@ public class Graph {
 	public void setEdgeList(List<Edge> edgeList) {
 		this.edgeList = edgeList;
 	}
-	public Map<BigInteger, Integer> getDegree() {
+	public Map<Integer, Integer> getDegree() {
 		return degree;
 	}
-	public void setDegree(Map<BigInteger, Integer> degree) {
+	public void setDegree(Map<Integer, Integer> degree) {
 		this.degree = degree;
 	}
